move label_translation into router and rename to RoutingTable
