+++ tests/unit/test_router.py
@@ -2,26 +2,23 @@
 #  See LICENSE file for licensing details.
 
 """Unit tests for the router module."""
+import itertools
 from unittest.mock import MagicMock
 
 import pytest
 
 from webhook_router.mq import add_job_to_queue
-from webhook_router.router import RouterError, forward
-from webhook_router.webhook import Job, JobStatus
-from webhook_router.webhook.label_translation import (
-    InvalidLabelCombinationError,
-    LabelsFlavorMapping,
-    labels_to_flavor,
+from webhook_router.router import (
+    LABEL_SEPARATOR,
+    FlavorLabelsMapping,
+    RouterError,
+    RoutingTable,
+    _InvalidLabelCombinationError,
+    _labels_to_flavor,
+    forward,
+    to_routing_table,
 )
-
-
-@pytest.fixture(name="labels_to_flavor_mock")
-def labels_to_flavor_mock_fixture(monkeypatch: pytest.MonkeyPatch) -> MagicMock:
-    """Mock the labels_to_flavor function."""
-    mock = MagicMock(spec=labels_to_flavor)
-    monkeypatch.setattr("webhook_router.router.labels_to_flavor", mock)
-    return mock
+from webhook_router.webhook import Job, JobStatus
 
 
 @pytest.fixture(name="add_job_to_queue_mock")
@@ -32,7 +29,6 @@ def add_job_to_queue_mock_fixture(monkeypatch: pytest.MonkeyPatch) -> MagicMock:
     return mock
 
 
-@pytest.mark.usefixtures("labels_to_flavor_mock")
 @pytest.mark.parametrize(
     "job_status, is_forwarded",
     [
@@ -55,25 +51,26 @@ def test_job_is_forwarded(
     # mypy does not understand that we can pass strings instead of HttpUrl objects
     # because of the underlying pydantic magic
     job = Job(
-        labels=["self-hosted", "linux", "arm64"],
+        labels=["arm64"],
         status=job_status,
         run_url="https://api.github.com/repos/f/actions/runs/8200803099",  # type: ignore
     )
     forward(
         job,
-        route_table=LabelsFlavorMapping(mapping={}, default_flavor="default", ignore_labels=set()),
+        routing_table=RoutingTable(
+            mapping={"arm64": "arm64"}, default_flavor="arm64", ignore_labels=set()
+        ),
     )
 
     assert add_job_to_queue_mock.called == is_forwarded
 
 
-def test_invalid_label_combination(labels_to_flavor_mock: MagicMock):
+def test_invalid_label_combination():
     """
     arrange: A job with an invalid label combination.
     act: Forward the job to the message queue.
     assert: A RouterError is raised.
     """
-    labels_to_flavor_mock.side_effect = InvalidLabelCombinationError("Invalid label combination")
     # mypy does not understand that we can pass strings instead of HttpUrl objects
     # because of the underlying pydantic magic
     job = Job(
@@ -84,8 +81,210 @@ def test_invalid_label_combination(labels_to_flavor_mock: MagicMock):
     with pytest.raises(RouterError) as e:
         forward(
             job,
-            route_table=LabelsFlavorMapping(
-                mapping={}, default_flavor="default", ignore_labels=set()
-            ),
+            routing_table=RoutingTable(mapping={}, default_flavor="default", ignore_labels=set()),
         )
-    assert str(e.value) == "Not able to forward job: Invalid label combination"
+    assert "Not able to forward job: Invalid label combination:" in str(e.value)
+
+
+def test_to_routing_table():
+    """
+    arrange: A mapping of flavors to labels.
+    act: Call to_routing_table with the mapping.
+    assert: A LabelsFlavorMapping object is returned.
+    """
+    flavor_mapping = FlavorLabelsMapping(
+        mapping=[("large", ["arm64", "large"]), ("x64-large", ["large", "x64", "jammy"])]
+    )
+    ignore_labels = {"self-hosted", "linux"}
+    labels_mapping = to_routing_table(flavor_mapping, ignore_labels)
+    assert labels_mapping == RoutingTable(
+        mapping={
+            "arm64": "large",
+            "large": "large",
+            f"arm64{LABEL_SEPARATOR}large": "large",
+            "x64": "x64-large",
+            "jammy": "x64-large",
+            f"jammy{LABEL_SEPARATOR}x64": "x64-large",
+            f"jammy{LABEL_SEPARATOR}large": "x64-large",
+            f"large{LABEL_SEPARATOR}x64": "x64-large",
+            f"jammy{LABEL_SEPARATOR}large{LABEL_SEPARATOR}x64": "x64-large",
+        },
+        default_flavor="large",
+        ignore_labels=ignore_labels,
+    )
+
+
+def test_to_routing_table_case_insensitive():
+    """
+    arrange: A mapping of flavors to labels with labels in mixed case.
+    act: Call to_routing_table with the mapping.
+    assert: A LabelsFlavorMapping object is returned which has all labels in lower case.
+    """
+    flavor_mapping = FlavorLabelsMapping(mapping=[("large", ["arM64", "LaRgE"])])
+    ignore_labels = {"self-HOSTED", "LINux"}
+    labels_mapping = to_routing_table(flavor_mapping, ignore_labels)
+    assert labels_mapping == RoutingTable(
+        mapping={
+            "arm64": "large",
+            "large": "large",
+            f"arm64{LABEL_SEPARATOR}large": "large",
+        },
+        default_flavor="large",
+        ignore_labels={"self-hosted", "linux"},
+    )
+
+
+def test_labels_to_flavor():
+    """
+    arrange: Two flavors and a labels to flavor mapping
+    act: Call labels_to_flavor with all combination of labels.
+    assert: The correct flavor is returned.
+    """
+    arm_flavor_labels = ["arm64", "jammy", "large"]
+    x64_flavor_labels = ["large", "noble", "x64"]
+    arm_label_combination = set(
+        x
+        for length in range(1, len(arm_flavor_labels) + 1)
+        for x in itertools.combinations(arm_flavor_labels, length)
+    )
+    x64_label_combination = set(
+        x
+        for length in range(1, len(x64_flavor_labels) + 1)
+        for x in itertools.combinations(x64_flavor_labels, length)
+    )
+    x64_label_combination.remove(("large",))
+    mapping = RoutingTable(
+        mapping={
+            **{
+                LABEL_SEPARATOR.join(label_combination): "large"
+                for label_combination in arm_label_combination
+            },
+            **{
+                LABEL_SEPARATOR.join(label_combination): "x64-large"
+                for label_combination in x64_label_combination
+            },
+        },
+        default_flavor="large",
+        ignore_labels={"self-hosted", "linux"},
+    )
+
+    for label_combination in arm_label_combination:
+        assert (
+            _labels_to_flavor(set(label_combination), mapping) == "large"
+        ), f"Expected large flavor for {label_combination}"
+
+    for label_combination in x64_label_combination:
+        assert (
+            _labels_to_flavor(set(label_combination), mapping) == "x64-large"
+        ), f"Expected x64-large flavor for {label_combination}"
+
+
+def test_labels_to_flavor_case_insensitive():
+    """
+    arrange: Two flavors and a labels to flavor mapping
+    act: Call labels_to_flavor with all combination of labels in mixed case.
+    assert: The correct flavor is returned.
+    """
+    mapping = RoutingTable(
+        mapping={
+            "small": "small",
+            "arm64": "large",
+            "large": "large",
+            f"arm64{LABEL_SEPARATOR}large": "large",
+        },
+        default_flavor="large",
+        ignore_labels={"self-hosted", "linux"},
+    )
+
+    assert _labels_to_flavor({"SMALl"}, mapping) == "small"
+    assert _labels_to_flavor({"ARM64"}, mapping) == "large"
+    assert _labels_to_flavor({"lARGE"}, mapping) == "large"
+    assert _labels_to_flavor({"arM64", "lArge"}, mapping) == "large"
+    assert _labels_to_flavor({"small", "SELF-hosted"}, mapping) == "small"
+
+
+def test_labels_to_flavor_default_label():
+    """
+    arrange: Two flavors and a labels to flavor mapping
+    act: Call labels_to_flavor with empty labels.
+    assert: The default flavor is returned.
+    """
+    mapping = RoutingTable(
+        mapping={
+            "arm64": "large",
+            "large": "large",
+            f"arm64{LABEL_SEPARATOR}large": "large",
+            "x64": "large-x64",
+            f"large{LABEL_SEPARATOR}x64": "x64-large",
+        },
+        default_flavor="large",
+        ignore_labels={"self-hosted", "linux"},
+    )
+    assert _labels_to_flavor(set(), mapping) == "large"
+
+
+def test_labels_to_flavor_invalid_combination():
+    """
+    arrange: Two flavors and a labels to flavor mapping
+    act: Call labels_to_flavor with an invalid combination.
+    assert: An InvalidLabelCombinationError is raised.
+    """
+    mapping = RoutingTable(
+        mapping={
+            "arm64": "large",
+            "large": "large",
+            f"arm64{LABEL_SEPARATOR}large": "large",
+            "x64": "large-x64",
+            f"large{LABEL_SEPARATOR}x64": "x64-large",
+        },
+        default_flavor="large",
+        ignore_labels={"self-hosted", "linux"},
+    )
+    labels = {"self-hosted", "linux", "arm64", "large", "x64"}
+    with pytest.raises(_InvalidLabelCombinationError) as exc_info:
+        _labels_to_flavor(labels, mapping)
+    assert "Invalid label combination:" in str(exc_info.value)
+
+
+def test_labels_to_flavor_unrecognised_label():
+    """
+    arrange: Two flavors and a labels to flavor mapping
+    act: Call labels_to_flavor with an unrecognised label.
+    assert: An InvalidLabelCombinationError is raised.
+    """
+    mapping = RoutingTable(
+        mapping={
+            "arm64": "large",
+            "large": "large",
+            f"arm64{LABEL_SEPARATOR}large": "large",
+            "x64": "large-x64",
+            f"large{LABEL_SEPARATOR}x64": "x64-large",
+        },
+        default_flavor="large",
+        ignore_labels={"self-hosted", "linux"},
+    )
+    labels = {"self-hosted", "linux", "arm64", "large", "noble"}
+    with pytest.raises(_InvalidLabelCombinationError) as exc_info:
+        _labels_to_flavor(labels, mapping)
+    assert "Invalid label combination:" in str(exc_info.value)
+
+
+def test_labels_to_flavor_ignore_labels():
+    """
+    arrange: Two flavors and a labels to flavor mapping
+    act: Call labels_to_flavor with an ignored label.
+    assert: The correct flavor is returned.
+    """
+    mapping = RoutingTable(
+        mapping={
+            "arm64": "large",
+            "large": "large",
+            f"arm64{LABEL_SEPARATOR}large": "large",
+            "x64": "large-x64",
+            f"large{LABEL_SEPARATOR}x64": "x64-large",
+        },
+        default_flavor="large",
+        ignore_labels={"self-hosted", "linux"},
+    )
+    labels = {"self-hosted", "linux", "arm64", "large"}
+    assert _labels_to_flavor(labels, mapping) == "large"
